python/requirements.txt
Flask
Flask-SQLAlchemy
marshmallow
psycopg2-binary
